// src/logistics-api/model/updateCargoByLogistician.ts
/**
 * Logistics API
 * Logistics API provides services to find and manage entities for logistics
 *
 * OpenAPI spec version: 1.0
 * Contact: vladislavprokopenko1@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Object representing Cargo updating by Logistician
 */
export interface UpdateCargoByLogistician { 
    cargoName?: string;
    /**
     * Cargo weight, tons
     */
    weight?: number;
    startCity?: string;
    startState?: string;
    startAddress?: string;
    finalCity?: string;
    finalState?: string;
    finalAddress?: string;
}